Added explicit cast of val_ to long in VarCHRM::put(double val_) in the case where kind is FUNobs before running functions ReadAheadObsMacro or WriteAheadObsMacro.

+++ release1.1/src/core/CRHM_parse.cpp
@@ -89,9 +89,9 @@ void VarCHRM::put(double val_) {
 		break;
 	case FUNobs:
 		if (name == "ReadAheadObs")
-			val = !Me->ReadAheadObsMacro(val_);
+			val = !Me->ReadAheadObsMacro((long) val_);
 		else
-			val = !Me->WriteAheadObsMacro(val_);
+			val = !Me->WriteAheadObsMacro((long) val_);
 		break;
 	default:
 		break;
